Moved create_buffer and create_vertex buffer to new submodule
